kube: update cert-manager
And update the manifest to use the latest crd definition.
Now the certificate issuer need to be on the same namespace
with the ingress as it can't use the secret from outside it's own
namespace.

While at it remove wildcard certs issuer as we only need 1 certificate
for rndir.cc