deploy: add cronjob to kill Shutdown pods
GKE now introduce graceful node shutdown and this caused pods on this
node to be evicted and marked as failed.

From their docs this pods will be eventually deleted once it hits the
treshold, but the problem is this values is too high (12500) and not
configurable from our side.

To avoid this, wo we are going to kill this pods ourself by adding a
cronjob that will query these pods using kubectl commands.